# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for generating traces and retrieving source code mappings for performance analysis",
        "title": "Tempo OTLP Trace Demo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.4.208:3202",
    "basePath": "/",
    "paths": {
...
        "/api/report/generate": {
            "post": {
                "description": "Generates a report with comprehensive tracing. LONG TRACE - Generates 10-12 spans with 1500-3500ms duration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Generate a report",
                "parameters": [
                    {
                        "description": "Report generation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report generated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user/profile": {
            "get": {
                "description": "Retrieves user profile information. Generates 4-5 spans with 110-310ms duration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (default: user_12345)",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.UserProfileResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ReportRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "report_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "models.ReportResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "report_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.UserProfileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preferences": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}